backend/requirements.txt
aiofiles==25.1.0
fastapi==0.104.1
google-genai==1.52.0
matplotlib==3.9.2
numpy==1.24.3
pandas==2.2.2
plotly==6.0.1
pydantic==2.10.6
python-dotenv==1.0.1
python-multipart==0.0.6
requests==2.32.5
seaborn==0.13.2
tabulate==0.9.0
uvicorn==0.24.0
